testTask/requirements.txt
Django==2.2.12
djangorestframework==3.11.0
psycopg2-binary==2.8.5
psycopg2==2.7.4
django-environ==0.4.5

